pyproject.toml
[project]
name = "dataloader"
version = "0.0.0b4"
description = "Recipe-driven data loading framework for declarative EL workflows"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "json-log-formatter>=0.5.0",
    "pyarrow>=14.0",
    "fsspec>=2023.1.0",
]
authors = [
    {name = "DataLoader Contributors"},
]
keywords = ["etl", "data-loading", "data-pipeline", "recipe", "yaml"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.optional-dependencies]
postgres = [
    "psycopg2-binary>=2.9",
    "sqlalchemy>=2.0.0",
    "pandas>=2.0",
]
s3 = [
    "boto3>=1.28",
    "s3fs>=2023.1.0",
]
duckdb = [
    "duckdb>=0.9",
]
parquet = [
    "pandas>=2.0",
]
all = [
    "psycopg2-binary>=2.9",
    "sqlalchemy>=2.0.0",
    "pandas>=2.0",
    "boto3>=1.28",
    "s3fs>=2023.1.0",
    "duckdb>=0.9",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "moto>=4.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "isort>=6.1.0",
]

[project.scripts]
dataloader = "dataloader.cli.main:main"
